+++ src/main/java/com/phantomwing/rusticdelight/item/custom/ChocolateCoffeeItem.java
@@ -1,9 +1,11 @@
-package com.phantomwing.rusticdelight.block.custom;
+package com.phantomwing.rusticdelight.item.custom;
 
 import net.minecraft.core.Holder;
 import net.minecraft.world.effect.MobEffect;
+import net.minecraft.world.effect.MobEffectCategory;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.neoforged.neoforge.common.EffectCures;
@@ -12,9 +14,9 @@ import vectorwing.farmersdelight.common.item.DrinkableItem;
 import java.util.ArrayList;
 import java.util.Iterator;
 
-public class MilkCoffeeItem extends DrinkableItem
+public class ChocolateCoffeeItem extends DrinkableItem
 {
-    public MilkCoffeeItem(Properties properties) {
+    public ChocolateCoffeeItem(Item.Properties properties) {
         super(properties, true, true);
     }
 
@@ -25,7 +27,7 @@ public class MilkCoffeeItem extends DrinkableItem
 
         while (itr.hasNext()) {
             MobEffectInstance effect = itr.next();
-            if (effect.getCures().contains(EffectCures.MILK)) {
+            if (effect.getEffect().value().getCategory().equals(MobEffectCategory.HARMFUL) && effect.getCures().contains(EffectCures.MILK)) {
                 compatibleEffects.add(effect.getEffect());
             }
         }